supply acpl_id to h5Acreate2 to conform to API
An error in the HDF5 docs led to an error in the implementation, the docs described the acpl_id paramenter as unused
but actually it was only aapl_id that is (still) unused. The issue does not manifest if you only read/write with
the hdf5-rust API, because it always assumbed UTF-8 encoding which is essentially a superset of ASCII. However, another
API might not assume UTF-8 encoding if the default (ASCII) encoding is set.